Improves LCP by preloading avatar
Prioritizes avatar loading by preloading it as the LCP image.

This change moves the preload directives to the top of the
`<head>` section to ensure that the browser discovers them as
early as possible. It also updates the `srcset` and `sizes`
attributes of the avatar image for better responsiveness and
performance. Finally, it adds `decoding="async"` attribute to the image.